+++ src/main/java/com/configcat/override/LocalMapDataSource.java
@@ -1,4 +1,4 @@
-package com.configcat;
+package com.configcat.override;
 
 import com.configcat.models.Setting;
 import com.configcat.models.SettingType;
