+++ src/test/java/com/configcat/LocalTest.java
@@ -1,5 +1,7 @@
 package com.configcat;
 
+import com.configcat.override.OverrideBehaviour;
+import com.configcat.override.OverrideDataSource;
 import com.configcat.polling.PollingModes;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
